Made numerous grammatical fixes to the docs.
Also changed an image on the Nine-patch page to be less painful to look at.

Looks like most (all?) links have been lost in the reST conversion - I've added a couple of external ones back in that I happened to come across.
Inline code also seems to not be marked up as such - I've added backticks around the instances that I came across.